requirements.txt
azureml-sdk>=1.45.0
pytest>=7.0.0
flake8>=5.0.0
black>=24.0.0
isort>=5.0.0
detect-secrets>=1.4.0
pandas>=2.0.0
scikit-learn>=1.3.0
